content-security-policy/svg/svg-inline.html

<!DOCTYPE html>
<html>
<head>
    <meta http-equiv="Content-Security-Policy" content="script-src 'self' 'unsafe-inline'"
    <title>svg-policy-with-resource</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
</head>

<body>
    <p>Tests that an SVG loaded inline enforces the loading context's CSP.</p>
    <div id="log"></div>
    <?xml version="1.0" standalone="no"?>
<svg width="6cm" height="5cm" viewBox="0 0 600 500"
     xmlns="http://www.w3.org/2000/svg" version="1.1">

    <script>
        var evalRan = false;
    </script>

   <script type="application/ecmascript"> <![CDATA[
    test(function() {assert_throws(new EvalError(), function() { eval('evalRan = true;') })}, "eval() should throw without 'unsafe-eval' keyword source in script-src directive.");
  ]]> </script>

  <!-- Act on each click event -->
  <circle onclick="circle_click(evt)" cx="300" cy="225" r="100"
          fill="lawngreen"/>

  <text x="300" y="250"
        font-family="Verdana"
        font-size="50"
        text-anchor="middle">
      PASS
    </text>
</svg>

    <script>
                test(function() {assert_false(evalRan);}, "eval() didn't run")
    </script>
</body>
</html>
